[fix]goodとbad保存時のdestroy、ifで条件つけた

+++ app/controllers/bads_controller.rb
@@ -3,7 +3,9 @@ class BadsController < ApplicationController
     bad = current_user.bads.build(review_id: params[:review_id])
     bad.save
     good = Good.find_by(user_id: current_user.id, review_id: params[:review_id])
-    good.destroy
+    if good.present?
+      good.destroy
+    end
     redirect_to request.referer
   end
