+++ app/controllers/goods_controller.rb
@@ -3,7 +3,9 @@ class GoodsController < ApplicationController
     good = current_user.goods.build(review_id: params[:review_id])
     good.save
     bad = Bad.find_by(user_id: current_user.id, review_id: params[:review_id])
-    bad.destroy
+    if bad.present?
+      bad.destroy
+    end
     redirect_to request.referer
   end
